+++ src/main/java/com/vapeshop/config/DBConnect.java
@@ -1,9 +1,9 @@
 package com.vapeshop.config;
 
 
-import java.sql.*;
-import java.util.logging.Level;
-import java.util.logging.Logger;
+import java.sql.Connection;
+import java.sql.DriverManager;
+import java.sql.SQLException;
 
 public class DBConnect {
     public static String dbName = "VapeShopDB";
