+++ src/main/java/com/vapeshop/controller/employee/product/ProductTypeCreateServlet.java
@@ -1,12 +1,15 @@
 package com.vapeshop.controller.employee.product;
 
 import com.vapeshop.entity.ImageProduct;
-import com.vapeshop.entity.Product;
 import com.vapeshop.entity.ProductType;
 import com.vapeshop.respository.employee.ProductRespository;
-import jakarta.servlet.*;
-import jakarta.servlet.http.*;
-import jakarta.servlet.annotation.*;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.annotation.MultipartConfig;
+import jakarta.servlet.annotation.WebServlet;
+import jakarta.servlet.http.HttpServlet;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import jakarta.servlet.http.Part;
 
 import java.io.File;
 import java.io.IOException;
