+++ src/main/java/com/vapeshop/controller/employee/product/ProductUpdateServlet.java
@@ -2,9 +2,11 @@ package com.vapeshop.controller.employee.product;
 
 import com.vapeshop.entity.Product;
 import com.vapeshop.respository.employee.ProductRespository;
-import jakarta.servlet.*;
-import jakarta.servlet.http.*;
-import jakarta.servlet.annotation.*;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.annotation.WebServlet;
+import jakarta.servlet.http.HttpServlet;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 
 import java.io.IOException;
