+++ src/main/java/com/vapeshop/controller/login/ForgotPasswordServlet.java
@@ -88,5 +88,8 @@ public class ForgotPasswordServlet extends HttpServlet {
         }
 
     }
-
+    @Override
+    protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+     request.getRequestDispatcher("forgotPassword.jsp").forward(request,response);
+    }
 }
