+++ src/main/java/com/vapeshop/controller/login/LoginServlet.java
@@ -85,7 +85,7 @@ public class LoginServlet extends HttpServlet {
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
-        processRequest(request, response);
+          request.getRequestDispatcher("login.jsp").forward(request,response);
     }
 
     /**
