+++ src/main/java/com/vapeshop/controller/login/RegisterServlet.java
@@ -114,7 +114,7 @@ public class RegisterServlet extends HttpServlet {
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
-        processRequest(request, response);
+       request.getRequestDispatcher("register.jsp").forward(request,response);
     }
 
     /**
