--- src/main/java/com/vapeshop/khanh.java
@@ -1,5 +1,0 @@
-package com.vapeshop;
-
-public class khanh {
-
-}
