+++ src/main/java/com/vapeshop/respository/ProductRepository.java
@@ -10,7 +10,6 @@ import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.ArrayList;
-import java.util.List;
 
 public class ProductRepository {
     //Get all Product from Product,productType,ImageProduct table
